requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
coverage==4.5.4
Django==2.2.4
django-cors-headers==3.1.0
django-dotenv==1.4.2
django-filter==2.2.0
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
Faker==2.0.0
gunicorn==19.9.0
idna==2.8
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
